Move Auth classes to "web" package. RDFConneection support LibSec.
Rename AuthMode->AuthScheme
